feat:placed trigger in assetview
